% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{mzTest}
\alias{mzTest}
\title{Maller-Zhou test}
\usage{
mzTest(time, status)
}
\arguments{
\item{time}{A numeric vector for the observed survival times.}

\item{status}{A numeric vector for the event indicator;
0 indicates right-censoring and 1 indicates events.}
}
\description{
Performs the Maller-Zhou test.
}
\examples{
data(Teeth500)
mzTest(Teeth500$time, Teeth500$event)
}
