% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geelm.R
\name{plot.geelm}
\alias{plot.geelm}
\title{Plot method for 'geelm' objects}
\usage{
\method{plot}{geelm}(x, type = c("residuals", "cv", "trace"), ...)
}
\arguments{
\item{x}{An object of class 'pCure', usually returned by the 'pCure()' function.}

\item{type}{A character string specifying the type of plot to generate.
Available options are "residuals," "cv," and "trace,"
which correspond to the pseudo-residual plot, cross-validation plot,
and trace plot for different values of the tuning parameter, respectively.}

\item{...}{Other arguments for future extension.}
}
\value{
A ggplot object representing the residual plot, cross-validation plot,
or the trace plot for an object of class \code{"geelm"}. 
This can be further modified using \code{"ggplot2"} functions.
}
\description{
Plot method for 'geelm' objects
}
