% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoCure_pkg.R
\docType{package}
\name{pseudoCure-package}
\alias{pseudoCure}
\alias{pseudoCure-package}
\alias{pseudoCure-packages}
\title{pseudoCure: A pseudo-observations approach for analyzing survival data with a cure fraction}
\description{
A collection of easy-to-use tools for regression analysis of survival data with a cure fraction.
The modeling framework is based on the Cox proportional hazards mixture cure model and
the bounded cumulative hazard model.
The pseudo-observations approach is utilized to assess covariate effects and
embedded in the variable selection procedure.
}
\author{
\strong{Maintainer}: Sy Han (Steven) Chiou \email{schiou@smu.edu}

Authors:
\itemize{
  \item Chien-Lin Su \email{marksu740824@gmail.com}
  \item Feng-Chang Lin \email{flin@bios.unc.edu}
}

}
