% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCure.R
\name{pCure}
\alias{pCure}
\title{Cure Rate Model with pseudo-observation approach}
\usage{
pCure(
  formula1,
  formula2,
  time,
  status,
  data,
  subset,
  t0,
  model = c("mixture", "promotion"),
  nfolds = 5,
  lambda1 = NULL,
  exclude1 = NULL,
  penalty1 = c("scad", "lasso"),
  lambda2 = NULL,
  exclude2 = NULL,
  penalty2 = c("scad", "lasso"),
  control = list()
)
}
\arguments{
\item{formula1}{A formula object starting with \code{~} for the model formula.
This specifies the covariates in the incidence component and the long-term component under
the mixture cure model and the bounded cumulative model, respectively.}

\item{formula2}{A formula object starting with \code{~} for the model formula.
This specifies the covariates in the latency component and the short-term component under
the mixture cure model and the bounded cumulative model, respectively.}

\item{time}{A numeric vector for the observed survival times.}

\item{status}{A numeric vector for the event indicator;
0 indicates right-censoring and 1 indicates events.}

\item{data}{An optional data frame that contains the covariates and response variables
(\code{time} and \code{event}).}

\item{subset}{An optional logical vector specifying a subset of
observations to be used in the fitting process.}

\item{t0}{A vector of times, where the pseudo-observations are constructed.
When not specified, the default values are the 10, 20, ..., 90, 95 percentiles of
uncensored event times.}

\item{model}{A character string specifying the underlying model.
The available functional form are \code{"mixture"} and \code{"promotion"}
correspond to the mixture cure model and the bounded cumulative model, respectively.}

\item{nfolds}{An optional integer value specifying the number of folds.
The default value is 5.}

\item{lambda1, lambda2}{An option for specifying the tuning parameter used in penalization.
When this is unspecified or has a \code{NULL} value,
penalization will not be applied and \code{pCure()} will uses all covariates
specified in the formulas.
Alternatively, this can be specified as a vector numeric vector of non-negative values
or "auto" for auto selection.}

\item{exclude1, exclude2}{A character string specifying variables to
exclude in variable selection.}

\item{penalty1, penalty2}{A character string specifying the penalty function.
The available options are code{"scad"} and code{"lasso"}.}

\item{control}{A list of control parameters. See detail.}
}
\description{
Fits either a mixture cure model or a bounded cumulative hazard (promotion time) model
with pseudo-observation approach.
}
\examples{
## A melanoma data example
data(e1684, package = "smcure")
e1684 <- na.omit(e1684)

## Define formula & fix random seed 
fn <- ~ AGE + TRT * SEX; set.seed(123)

## Unpenalized PHMC model (~ 0.046s)
summary(fit1 <- pCure(fn, fn, FAILTIME, FAILCENS, e1684))

## Penalized PHMC model with tuning parameters selected by 10-fold cross validation
## User specifies the range of tuning parameters
summary(update(fit1, lambda1 = 1:10 / 200, lambda2 = 1:10 / 200))
## Auto selection of the range of tuning parameters
summary(update(fit1, lambda1 = "auto", lambda2 = "auto"))

## Penalized PHMC model given tuning parameters
summary(update(fit1, lambda1 = 0.006, lambda2 = 0.022))
}
\references{
Su, C.-L., Chiou, S., Lin, F.-C., and Platt, R. W. (2022)
Analysis of survival data with cure fraction and variable selection: A pseudo-observations approach
\emph{Statistical Methods in Medical Research}, \bold{31}(11): 2037--2053.
}
